+++ otx/algorithms/common/adapters/torch/utils/__init__.py
@@ -3,6 +3,6 @@
 # Copyright (C) 2023 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from .bs_search_algo import adapt_batch_size
+from .bs_search_algo import BsSearchAlgo
 
-__all__ = ["adapt_batch_size"]
+__all__ = ["BsSearchAlgo"]