+++ tests/integration/cli/classification/test_classification.py
@@ -266,11 +266,12 @@ class TestMultiClassClassificationCLI:
 
     @e2e_pytest_component
     @pytest.mark.parametrize("template", default_templates, ids=default_templates_ids)
-    def test_otx_train_auto_decrease_batch_size(self, template, tmp_dir_path):
-        decrease_bs_args = copy.deepcopy(args)
-        decrease_bs_args["train_params"].extend(["--learning_parameters.auto_decrease_batch_size", "true"])
-        tmp_dir_path = tmp_dir_path / "multi_class_cls_auto_decrease_batch_size"
-        otx_train_testing(template, tmp_dir_path, otx_dir, decrease_bs_args)
+    @pytest.mark.parametrize("bs_adapt_type", ["Safe", "Full"])
+    def test_otx_train_auto_adapt_batch_size(self, template, tmp_dir_path, bs_adapt_type):
+        adapting_bs_args = copy.deepcopy(args)
+        adapting_bs_args["train_params"].extend(["--learning_parameters.auto_adapt_batch_size", bs_adapt_type])
+        tmp_dir_path = tmp_dir_path / f"multi_class_cls_auto_adapt_{bs_adapt_type}_batch_size"
+        otx_train_testing(template, tmp_dir_path, otx_dir, adapting_bs_args)
 
 
 # Multi-label training w/ 'car', 'tree', 'bug' classes